Split functions.php into domain-specific files.
Organize ~1700 lines of functions into smaller, maintainable files:
- functions-activity.php: Activity type checking
- functions-comment.php: Comment handling and reply intent
- functions-federation.php: Federation, outbox, follow operations
- functions-post.php: Post operations and summary generation
- functions-request.php: HTTP requests and remote metadata
- functions-url.php: URL utilities and normalization
- functions-user.php: User/actor functions

Also restructure tests to match the new file organization.
