+++ tests/phpunit/tests/includes/functions/class-test-functions-comment.php
@@ -1,0 +1,167 @@
+<?php
+/**
+ * Test file for Comment Functions.
+ *
+ * @package Activitypub
+ */
+
+namespace Activitypub\Tests;
+
+/**
+ * Test class for Comment Functions.
+ *
+ * @coversDefaultClass \Activitypub
+ */
+class Test_Functions_Comment extends \WP_UnitTestCase {
+
+	/**
+	 * Post ID.
+	 *
+	 * @var int
+	 */
+	public $post_id;
+
+	/**
+	 * Set up the test.
+	 */
+	public function set_up() {
+		parent::set_up();
+
+		$this->post_id = \wp_insert_post(
+			array(
+				'post_author'  => 1,
+				'post_content' => 'test',
+			)
+		);
+	}
+
+	/**
+	 * Tear down.
+	 */
+	public function tear_down() {
+		parent::tear_down();
+
+		_delete_all_posts();
+	}
+
+	/**
+	 * Test object_id_to_comment.
+	 *
+	 * @covers \Activitypub\object_id_to_comment
+	 */
+	public function test_object_id_to_comment_basic() {
+		$single_comment_source_id = 'https://example.com/single';
+		$content                  = 'example comment that has bunch of text';
+		$comment_id               = \wp_new_comment(
+			array(
+				'comment_post_ID'      => $this->post_id,
+				'comment_author'       => 'Example User',
+				'comment_author_url'   => 'https://example.com/user',
+				'comment_content'      => $content,
+				'comment_type'         => '',
+				'comment_author_email' => '',
+				'comment_parent'       => 0,
+				'comment_meta'         => array(
+					'source_id'  => $single_comment_source_id,
+					'source_url' => 'https://example.com/123',
+					'avatar_url' => 'https://example.com/icon',
+					'protocol'   => 'activitypub',
+				),
+			),
+			true
+		);
+		$query_result             = \Activitypub\object_id_to_comment( $single_comment_source_id );
+		$this->assertInstanceOf( \WP_Comment::class, $query_result );
+		$this->assertEquals( $comment_id, $query_result->comment_ID );
+		$this->assertEquals( $content, $query_result->comment_content );
+	}
+
+	/**
+	 * Test object_id_to_comment with invalid source ID.
+	 *
+	 * @covers \Activitypub\object_id_to_comment
+	 */
+	public function test_object_id_to_comment_none() {
+		$single_comment_source_id = 'https://example.com/none';
+		$query_result             = \Activitypub\object_id_to_comment( $single_comment_source_id );
+		$this->assertFalse( $query_result );
+	}
+
+	/**
+	 * Test object_id_to_comment with duplicate source ID.
+	 *
+	 * @covers \Activitypub\object_id_to_comment
+	 */
+	public function test_object_id_to_comment_duplicate() {
+		$duplicate_comment_source_id = 'https://example.com/duplicate';
+
+		add_filter( 'duplicate_comment_id', '__return_zero', 99 );
+		add_filter( 'wp_is_comment_flood', '__return_false', 99 );
+		for ( $i = 0; $i < 2; ++$i ) {
+			\wp_new_comment(
+				array(
+					'comment_post_ID'      => $this->post_id,
+					'comment_author'       => 'Example User',
+					'comment_author_url'   => 'https://example.com/user',
+					'comment_content'      => 'example comment',
+					'comment_type'         => '',
+					'comment_author_email' => '',
+					'comment_parent'       => 0,
+					'comment_meta'         => array(
+						'source_id'  => $duplicate_comment_source_id,
+						'source_url' => 'https://example.com/123',
+						'avatar_url' => 'https://example.com/icon',
+						'protocol'   => 'activitypub',
+					),
+				),
+				true
+			);
+		}
+		remove_filter( 'duplicate_comment_id', '__return_zero', 99 );
+		remove_filter( 'wp_is_comment_flood', '__return_false', 99 );
+
+		$query_result = \Activitypub\object_id_to_comment( $duplicate_comment_source_id );
+		$this->assertInstanceOf( \WP_Comment::class, $query_result );
+	}
+
+	/**
+	 * Test get comment ancestors.
+	 *
+	 * @covers \Activitypub\get_comment_ancestors
+	 */
+	public function test_get_comment_ancestors() {
+		$comment_id = wp_insert_comment(
+			array(
+				'comment_type'         => 'comment',
+				'comment_content'      => 'This is a comment.',
+				'comment_author_url'   => 'https://example.com',
+				'comment_author_email' => '',
+				'comment_meta'         => array(
+					'protocol' => 'activitypub',
+				),
+			)
+		);
+
+		$this->assertEquals( array(), \Activitypub\get_comment_ancestors( $comment_id ) );
+
+		$comment_array = get_comment( $comment_id, ARRAY_A );
+
+		$parent_comment_id = wp_insert_comment(
+			array(
+				'comment_type'         => 'parent comment',
+				'comment_content'      => 'This is a parent comment.',
+				'comment_author_url'   => 'https://example.com',
+				'comment_author_email' => '',
+				'comment_meta'         => array(
+					'protocol' => 'activitypub',
+				),
+			)
+		);
+
+		$comment_array['comment_parent'] = $parent_comment_id;
+
+		wp_update_comment( $comment_array );
+
+		$this->assertEquals( array( $parent_comment_id ), \Activitypub\get_comment_ancestors( $comment_id ) );
+	}
+}
