+++ tests/phpunit/tests/includes/functions/class-test-functions-request.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Test file for Request Functions.
+ *
+ * @package Activitypub
+ */
+
+namespace Activitypub\Tests;
+
+/**
+ * Test class for Request Functions.
+ *
+ * @coversDefaultClass \Activitypub
+ */
+class Test_Functions_Request extends ActivityPub_TestCase_Cache_HTTP {
+
+	/**
+	 * Test the get_remote_metadata_by_actor function.
+	 *
+	 * @covers \Activitypub\get_remote_metadata_by_actor
+	 */
+	public function test_get_remote_metadata_by_actor() {
+		$metadata = \Activitypub\get_remote_metadata_by_actor( 'pfefferle@notiz.blog' );
+		$this->assertEquals( 'https://notiz.blog/author/matthias-pfefferle/', $metadata['url'] );
+		$this->assertEquals( 'pfefferle', $metadata['preferredUsername'] );
+		$this->assertEquals( 'Matthias Pfefferle', $metadata['name'] );
+	}
+}
